Actor in place, it's up to the Worker now. NB: not handling failures yet.

+++ app/models/TaskDetails.scala
@@ -1,10 +1,7 @@
 package models
 
-import play.api.libs.json.Json
-import play.api.libs.json.Writes
-
 final case class TaskDetails(
-    linesProcessed: Int,
+    linesProcessed: Long,
     avgLinesProcessed: Double,
     state: TaskState,
     result: Option[String]
